requirements.txt
transformers>=4.41.0
datasets>=2.19.0
torch>=2.3.0
accelerate>=0.29.2
scikit-learn>=1.5.0
tqdm>=4.66.4
sentencepiece>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.7.1
bitsandbytes>=0.43.0
peft>=0.5.0
protobuf>=3.20.3